多数据源配置（二）：Spring-data-jpa

+++ src/main/java/com/lkl/springboot/service/DepartmentService.java
@@ -4,7 +4,7 @@ import com.lkl.springboot.po.Department;
 
 public interface DepartmentService {
 
-	public Integer saveDepa(Department department); 
+	public Long saveDepa(Department department); 
 	
 	
 }
